refactor: extract store schema migrations
